Adding zip to sort out later (see extended desc)
This is a midway-point commit of all of the files that I have in overleaf at the moment. They are contained in this zip file. This is after performing some rework that Dr. Berleant suggested (it also contains a lot of typeface and grammatical updates). 

I'll actually extract the .tex files and update the `master` branch with the appropriate copies once the proposal is finished. 

The chapter 3 (methodology) section will also need to be re-worked in the final dissertation